refactor(frontend): extract ImportTransactionsPage to src/pages/ (Phase 3)
Part of the Phase 3 migration. Route shrinks from 714 to 5 lines.

- New src/pages/ImportTransactionsPage.tsx owns the review step, form
  state (react-hook-form with useFieldArray), selection helpers, bulk
  actions, and the import loop.
- UploadStep hoisted to src/components/transactions/import/UploadStep.tsx
  (§6 — never declare helpers inside a page body). Uses useAccounts(select)
  to build the own-account options slice inside TanStack Query.
- buildPayload, parsedRowToFormValues, and the CSV_COLUMNS reference
  moved to src/components/transactions/import/importPayload.ts so the
  page file doesn't carry unrelated pure helpers.
- The in-page useEffect that auto-redirects to /transactions after a
  fully successful import is extracted to
  src/hooks/import/useRedirectOnImportSuccess.ts (§4 — zero useEffect
  in pages).

The navigation blocker modal and the in-flight invalidate-all-
transactions call via useQueryClient stay as-is; they are not in the
known divergences list. Phase 6 handles the remaining `any` types in
ImportReviewRow and SplitPopover.
